Fix win32 build.
This fixes the binutils build with the x86_64-w64-mingw32 toolchain.
The reason for this is that this toolchain's <sys/stat.h> redefines
'stat' through a macro, as with:

  #define  stat  _stat64

Because bucomm.h didn't include <sys/stat.h>, the signature of
the functions that use 'struct stat' didn't match the one in the
implementation (which internally used 'struct _stat64'), resulting
in a build error.

Change-Id: I32de64af6c5f3e70344fa7c22b40d9a83695c7df